Create initial project structure
This is a Gradle project using the Kotlin buildscript.

Using the scala plugin as Gatling is written in Scala and it
should make it easier to interact with the Gatling codebase if
we keep it the same, although I would really like to write it in
Kotlin.

Git-Versioner for versioning.