Update action so it finishes
Uses the core components stats engine to log the response details, which Gatling
picks up to finish the simulation. Using the core components clock to do the
timings.

Adds NameGen extension to use Gatling to generate request names.

Next we're going to need to get access to a preconfigured AWS Lambda client to
actually perform the lambda invocation. I guess this is where we use the
protocol to configure the client and then pass extra details such as payload
into the action.