Implement AWS Lambda invoke action
This replaces the "do something" comment with an implementation of that "something".

This uses the AWS Lambda client to make an InvokeRequest with a payload.

This is currently hardcoded to a local instance of localstack, with a dummy
function name and payload.

Next will be implementing the LambdaProtocol to provide configuration values for
the lambda client, and then a way to provide parameters for the invoke action.