Add Lambda protocol for configuration
This adds a bunch of stuff required to configure the AWS Lambda client using
standard Gatling configuration. This adds a LambdaProtocol that can be configured
by providing the region, endpoint, and credentials.

Additionally adds other stuff like a ProtocolKey and Components which seems to make
it easier to access the protocol from my action builder.

Does not currently contain a custom DSL to make protocol construction easier, that
can come later.