Move name generation to action builder
Having name generation in the Action made it so every invocation generated a new
name. I am pretty sure this is not intended behaviour given the way other actions in
Gatling work. I believe each name should be for each Action as built, rather than
each call of each action. This means that requests get split by the specific action
being built, so all the invocations that are the same go into one bucket, and other
invocations go into other buckets.

This is easily remediated by moving name generation into the action builder rather
than the action.

Right now this is still using the generic "invoke" name, but once we pass through
the function to be called we can update it to reference that instead.