Allow protocol configuration to be optional
Previously you would be forced to provide configuration for the AwsClient that is going
to be created, which is fine for when you want to but considerably less fine when you
want to use auto discovery instead.

This changes the implementation of the protocol to have all optional fields, and then
when we are creating the client we check if the option is defined, and if so we updated
the client builder to supply the value.

You still need to provide 'None' for everything when supplying the protocol, but we can
update that to be better later by providing a ProtocolBuilder