Change action name back to invoke
One of the side effects of changing the function name to an expression is that
we're not able to resolve it in the action builder any more, as we don't have
access to the session.

I tried moving it back into the action too but it's out of scope of the session
there too.

The alternative is to not pass function name in as an Expression, but it feels like
a feed of function name to payload could be useful? Maybe.

So we're just using the genName to generate a name called `invoke` with an atomic
ID after it for when we have more than one action.