Add ability to supply invocation timeout override
The default override is not long enough for long-lived lambda functions.

This change allows clients to override the timeout to be longer (or
shorter).

The timeout is currently applied to both the apiCallTimeout and the
apiCallAttemptTimeout, which probably isn't the intended way for it
to work but worst case scenario I can just provide two new methods
and deprecate this one if it causes any problems.