Update to Gatling 3.5.0
This adds the Gatling 3.5.0 and the Scala 2.13.4 library dependencies so that
we can target these versions.

This is a breaking, non-backwards compatible change. This is probably fine as
Gatling don't support versions lower than the latest, so the impetus is to
migrate to the latest version when possible. That's currently not possible if
you're using this library.

It might be worth tying our release versions to the Gatling release versions
so that we can maintain compatibility. That could cause issues if we make
any changes ourselves though.