Tie version with Gatling
This adds the versioner configuration required to tie our version to the
version of Gatling that we're targeting. I might forgo using the versioner
plugin in the future and just setting a `version` variable that's used for
both Gatling and us, but this way feels more flexible for now (For example,
if we want to release 3.5.1.1 for a hotfix or something).