Remove git-versioner plugin
This removes the git-versioner plugin so that we can version manually.

I'm making this change as we want to track Gatling changes (So that our
release for Gatling x.y.z is version x.y.z) which is currently possible with
the versioner plugin, but if we add any extra functionality outside of a
Gatling release schedule, it becomes a bit strange to use. As in, if we
wanted to add functionality on top of Gatling 3.6.0, we'd release extension at
3.6.0.1, to make it clear it's still for Gatling 3.6.0. This isn't possible
using the versioner, as this would merge in as v3.6.0.2 at the minumum due to
the merge commit, and possibly more.

Now we'll have to remember to bump the version in the gradle file, which I
think will be easier than remembering to bump it in the merge commit message.

I still stand by the versioner plugin being a good idea (obviously), but just
not for this use case in particular.