Add ability to provide custom request name for invocations
This adds the ability to provide a custom name for the invocation, that will
show up in the Gatling report for that execution.

This should help identify different requests when making multiple invocations
in a chain- for example, if the lambda is invoked first with one payload, then
with another afterwards.

Fixes #17