Add basic configuration test
This adds a basic test that uses the expected configuration in the extension
and makes sure that the Gatling tests attempt to run.

They don't pass as they're not hooked up to a real AWS lambda, but I
might be able to add a proper localstack test at some point.